+++ lib/src/widgets/read_more_text.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
 
+/// TrimMode enum
 enum TrimMode {
   Length,
   Line,