+++ lib/src/widgets/restart_app_widget.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 
-/// RestartAppWidget
+/// RestartAppWidget helps you to restart your Flutter app
 class RestartAppWidget extends StatefulWidget {
   final Widget child;